+++ snuba/cli/devserver.py
@@ -24,7 +24,7 @@ def devserver(bootstrap, workers):
     daemons = [
         ('api', [
             'uwsgi', '--master', '--manage-script-name',
-            '--wsgi-file', 'snuba/api.py',
+            '--wsgi-file', 'snuba/views.py',
             '--http', '0.0.0.0:1218',
             '--http-keepalive', '--need-app', '--die-on-term',
         ]),
