+++ tests/base.py
@@ -213,7 +213,7 @@ class BaseEventsTest(BaseDatasetTest):
 class BaseApiTest(BaseEventsTest):
     def setup_method(self, test_method, dataset_name='events'):
         super().setup_method(test_method, dataset_name)
-        from snuba.api import application
+        from snuba.views import application
         assert application.testing is True
         application.config['PROPAGATE_EXCEPTIONS'] = False
         self.app = application.test_client()
