+++ tests/test_split.py
@@ -2,13 +2,13 @@ import pytest
 from typing import Any, Mapping
 
 from snuba import state
+from snuba.api.split import split_query
 from snuba.datasets import Dataset
 from snuba.datasets.factory import get_dataset
 from snuba.query.query import Query
-from snuba.query_engine import QueryResult
+from snuba.api.query import QueryResult
 from snuba.request import Request
 from snuba.request.request_settings import RequestSettings
-from snuba.split import split_query
 from snuba.utils.metrics.timer import Timer
