+++ tests/test_state.py
@@ -15,7 +15,7 @@ from snuba.state import safe_dumps
 class TestState(BaseEventsTest):
     def setup_method(self, test_method):
         super(TestState, self).setup_method(test_method)
-        from snuba.api import application
+        from snuba.views import application
         assert application.testing == True
         self.app = application.test_client()
         self.app.post = partial(self.app.post, headers={'referer': 'test'})
